Use tell() instead of calling open() twice.
We need to know whether the specfile is a new file (and thus needs a
header written at the top) or an existing file that we are appending to.
The exisitng approach does this by trying to open the file in 'x' mode
and, if that fails, opening the file in 'a' mode. The downside of this
approach is that the first failed attempt has a side-effect: the
filename is reserved. The MultiFileManager first reserves the name and
then tries to open the file, which seems like the correct order in
general, but means that we can't use this try-x-and-then-a approach.

I think this approach is better all around. I have tested that it works
on ordinary files and on `StringIO` buffers.