Accept stop documents with no 'reason'.
The 'reason' key is not required by the schema, and we found some
documents at CHX that do not have it.

It's a separate question how we are ending up with documents that have
an 'exit_status' other than 'success' by no 'reason'. The RunEngine
should be ensuring that. Nonetheless, this suitcase should be tolerant
of any document that meets the schema, so this change is correct.